refactor(pr-review): share mobile-safe context contracts
